added documentation and few annotations
